Typecasted hitCount to double while calculating hitRatio in LRUPolicy.cpp

+++ vmPrism/LRUPolicy.cpp
@@ -32,7 +32,7 @@ LRU::LRU(size_t cacheSize, std::vector<size_t> pageRefList)
 		}
 	}
 
-	hitRatio = hitCount / pageRefList.size();
+	hitRatio = ((double) hitCount) / pageRefList.size();
 	pageFaultCount = pageRefList.size() - hitCount;
 }
